use database instead plane text

+++ app.rb
@@ -6,21 +6,41 @@ require 'rack'
 require 'haml'
 require 'sass'
 
+require File.dirname(__FILE__) + '/models/statements'
+
 set :haml, { :attr_wrapper => '"',
              :escape_html => true }
 
 enable :sessions
 
 get '/' do
-  @log = File.open('log').readlines
+  statements = Statements.reverse_order(:created_at).
+    limit(15).all.map {|s| "<li>(#{s.time}) #{s.user}: #{s.text}</li>\n" }
+  @log = statements
   haml :index
 end
 
+get '/logout' do
+  session[:name] = nil
+  redirect '/'
+end
+
 post '/say' do
-  File.open('log', 'a') do |io|
-    io.puts "#{params['say']}<br>"
+  return unless request.xhr?
+
+  Statements.create do |s|
+    s.text = params['say']
+    s.user = session[:name]
+    s.created_at = Time.now
   end
   Process.kill :USR1, $$
+end
+
+post '/login' do
+  unless params['name'].empty?
+    session[:name] = params['name']
+  end
+  redirect '/'
 end
 
 get '/css/application.css' do
